move definitions of createSmallDeformationProcess() to cpp file
